Read addons into a list for more flexibility
Until now we used to scan directories every time
we wanted to either display the list of addons
or load addons in order to execute them. That
is wasteful and inflexible.

Use a list of elements and only update it when
needed, for example when the directory folder
notifies us of a change.

This is a stop towards disabling particular elements.