Undo tabs-in-processes, it didn't work out
We are not literally undoing in the sense of
reverting, this is actually about undoing
the individual parts of the last refactoring
that implemented the socket/ plug logic.

As turned out, the idea was nice but the
implementation was absolutely not, embedding
of processes in a graphical way is not
at all reliable enough for complex use cases.

Naturally this change should solve all sorts
of peculiar issues, including actual
regressions in functionality. Relieving.