Store history dates in 'token' to avoid locatime() and revamp deleting
Loading history items used to involve frequent calls to
localtime() and g_memdup(). We avoid this by storing the
date strings in parent items now, which can even be used to
display items in the treeview.
Deleting of treeview items was also reworked so it should
work from menu items and the panel.