Optimize location completion based on profiling
The test was modified slightly because we are not filtering
out trailing slahses anymore. We are not quite there still
when it comes to treating URIs as we should but for all
practical purposes it should be quite a bit faster.
The most annoying shortcoming still is that umlauts or any
special characters won't match at all.