Check for Maemo only if /etc/osso-af-init/osso-gtk.defs exists
A surprisingly high number of users mistakenly installed libhildon1
because they noticed that it wasn't found and later wondered where
the menubar was.

All Maemo releases have the file osso-gtk.defs, including Mer, so
we only check for libhildon1 and libosso if it is present.
If needed --enable-hildon still enforces its use.