Improve XBEL format compatibility and support bookmark sharing
Loading of XBEL needs to handle title and desc children of the root
element. This is equal to using the according properties.

MicroB uses metadata without an owner and uses children rather than
properties. This is against the spec but we want to handle that.

A key value leak was plugged.

If the bookmark file is absolute, we assume it is a foreign file
and we shouldn't overwrite it.