Implement MD5 and SHA1 Link Fingerprint support
If a URL has an embedded link fingerprint the checksum
and algorithm is shown in the download prompt. After a
file has been saved to disk a warning icon will be shown
if the checksum doesn't match and clicking the icon will
explain that the file is erroneous and not open it.

See http://mdhashtool.mozdev.org/lfinfo.html